requirements.txt
fastapi==0.112.0
psycopg2-binary==2.9.10
uvicorn==0.22.0
pydantic~=2.10.6
pytest~=8.3.5
pytest-mock==3.14.0
pytest-asyncio==0.25.3
pytest-cov==6.0.0
httpx==0.28.1
